+++ forui/lib/foundation.dart
@@ -13,7 +13,8 @@ export 'src/foundation/tappable/tappable_group.dart' hide GroupEntry, TappableGr
 export 'src/foundation/time.dart';
 export 'src/foundation/tween.dart';
 export 'src/foundation/typeahead_controller.dart';
-export 'src/foundation/portal/portal.dart';
-export 'src/foundation/portal/portal_constraints.dart' hide FixedConstraints;
-export 'src/foundation/portal/portal_overflow.dart';
-export 'src/foundation/portal/portal_spacing.dart';
+export 'src/foundation/overlay/overlay.dart';
+export 'src/foundation/overlay/portal.dart';
+export 'src/foundation/overlay/portal_constraints.dart' hide FixedConstraints;
+export 'src/foundation/overlay/portal_overflow.dart';
+export 'src/foundation/overlay/portal_spacing.dart';