+++ forui/lib/src/foundation/overlay/composited_portal.dart
@@ -1,37 +1,17 @@
-// ignore_for_file: prefer_asserts_with_message
-
 import 'package:flutter/foundation.dart';
 import 'package:flutter/rendering.dart';
-import 'package:flutter/scheduler.dart';
 import 'package:flutter/widgets.dart';
 
 import 'package:forui/forui.dart';
-import 'package:forui/src/foundation/portal/composited_child.dart';
-import 'package:forui/src/foundation/portal/layer.dart';
-import 'package:forui/src/foundation/portal/portal_constraints.dart';
+import 'package:forui/src/foundation/overlay/composited_child.dart';
+import 'package:forui/src/foundation/overlay/composited_overlay.dart';
+import 'package:forui/src/foundation/overlay/portal_constraints.dart';
 
 /// A [CompositedPortal] positions itself relative to a [CompositedChild].
 ///
-/// ## Implementation details:
-/// This class is a copy of [CompositedTransformFollower] with the following enhancements:
-/// * Receives the global position of a linked [CompositedChild].
+/// Unlike [CompositedOverlay], it additionally performs anchor alignment, overflow handling, and constraint switching.
 @internal
-class CompositedPortal extends SingleChildRenderObjectWidget {
-  /// The notifier that is updated whenever the linked [CompositedChild] changes its global position.
-  final FChangeNotifier notifier;
-
-  /// The link object that connects this [CompositedPortal] with a [CompositedChild]s.
-  final ChildLayerLink link;
-
-  /// Whether to show the widget's contents when there is no corresponding [CompositedChild] with the same [link].
-  ///
-  /// When the widget is linked, the child is positioned such that it has the same global position as the linked
-  /// [CompositedChild].
-  ///
-  /// When the widget is not linked, then: if [showWhenUnlinked] is true, the child is visible and not repositioned; if
-  /// it is false, then child is hidden.
-  final bool showWhenUnlinked;
-
+class CompositedPortal extends CompositedOverlay {
   /// The portal's constraints.
   final FPortalConstraints constraints;
 
@@ -58,8 +38,6 @@ class CompositedPortal extends SingleChildRenderObjectWidget {
   final Offset offset;
 
   const CompositedPortal({
-    required this.notifier,
-    required this.link,
     required this.constraints,
     required this.portalAnchor,
     required this.childAnchor,
@@ -67,7 +45,9 @@ class CompositedPortal extends SingleChildRenderObjectWidget {
     required this.spacing,
     required this.overflow,
     required this.offset,
-    this.showWhenUnlinked = false,
+    required super.notifier,
+    required super.link,
+    super.showWhenUnlinked,
     super.key,
     super.child,
   });
@@ -105,9 +85,6 @@ class CompositedPortal extends SingleChildRenderObjectWidget {
   void debugFillProperties(DiagnosticPropertiesBuilder properties) {
     super.debugFillProperties(properties);
     properties
-      ..add(DiagnosticsProperty('notifier', notifier))
-      ..add(DiagnosticsProperty('link', link))
-      ..add(DiagnosticsProperty('showWhenUnlinked', showWhenUnlinked))
       ..add(DiagnosticsProperty('constraints', constraints))
       ..add(DiagnosticsProperty('portalAnchor', portalAnchor))
       ..add(DiagnosticsProperty('childAnchor', childAnchor))
@@ -118,18 +95,9 @@ class CompositedPortal extends SingleChildRenderObjectWidget {
   }
 }
 
-/// ## Implementation details:
-/// This class is a copy of [RenderFollowerLayer] with the following differences/enhancements:
-/// * Repaints whenever the linked [CompositedChild]'s globalOffset changes.
-/// * Contains a [ChildLayerLink] instead of a [LayerLink].
-/// * Changes the [paintBounds] to be the entire viewport.
-/// * Shifts the portal if it overflows out of the viewport.
+/// Unlike [RenderOverlayLayer], this additionally performs anchor alignment, overflow handling, and constraint switching.
 @internal
-class RenderPortalLayer extends RenderProxyBox {
-  FChangeNotifier _notifier;
-  ChildLayerLink _link;
-  Size _viewSize;
-  bool _showWhenUnlinked;
+class RenderPortalLayer extends RenderOverlayLayer {
   FPortalConstraints _portalConstraints;
   Alignment _portalAnchor;
   Alignment _childAnchor;
@@ -139,10 +107,6 @@ class RenderPortalLayer extends RenderProxyBox {
   Offset _offset;
 
   RenderPortalLayer({
-    required FChangeNotifier notifier,
-    required ChildLayerLink link,
-    required Size viewSize,
-    required bool showWhenUnlinked,
     required FPortalConstraints portalConstraints,
     required Alignment portalAnchor,
     required Alignment childAnchor,
@@ -150,31 +114,18 @@ class RenderPortalLayer extends RenderProxyBox {
     required Offset spacing,
     required FPortalOverflow overflow,
     required Offset offset,
-
-    RenderBox? child,
-  }) : _notifier = notifier,
-       _link = link,
-       _viewSize = viewSize,
-       _showWhenUnlinked = showWhenUnlinked,
-       _portalConstraints = portalConstraints,
+    required super.notifier,
+    required super.link,
+    required super.viewSize,
+    required super.showWhenUnlinked,
+    super.child,
+  }) : _portalConstraints = portalConstraints,
        _portalAnchor = portalAnchor,
        _childAnchor = childAnchor,
        _padding = padding,
        _spacing = spacing,
        _overflow = overflow,
-       _offset = offset,
-       super(child);
-
-  @override
-  void attach(PipelineOwner owner) {
-    super.attach(owner);
-    notifier.addListener(_schedule);
-  }
-
-  /// Marks for repaint between frames so that calculations are performed using consistent state.
-  /// This was inspired by visibility_detector's implementation:
-  /// https://github.com/google/flutter.widgets/blob/494c6abd3de44a92d34c5cbc424db2eefe3915cf/packages/visibility_detector/lib/src/render_visibility_detector.dart#L46
-  void _schedule() => SchedulerBinding.instance.scheduleTask(markNeedsPaint, Priority.touch);
+       _offset = offset;
 
   @override
   void performLayout() {
@@ -203,7 +154,7 @@ class RenderPortalLayer extends RenderProxyBox {
   }
 
   @override
-  void paint(PaintingContext context, Offset offset) {
+  Offset onPaint(PaintingContext context, Offset offset) {
     assert(
       link.childSize != null || link.childLayer == null || childAnchor == .topLeft,
       '$link: layer is linked to ${link.childLayer} but a valid childSize is not set. '
@@ -211,8 +162,7 @@ class RenderPortalLayer extends RenderProxyBox {
       '(current value is $childAnchor).',
     );
 
-    final linkedOffset =
-        this.offset +
+    return this.offset +
         switch ((link.childRenderBox?.localToGlobal(.zero), link.childSize, child)) {
           (final childOffset?, final childSize?, final portal?) => overflow(
             // There is NO guarantee that this render box's size is the window's size. Always use viewSize.
@@ -227,134 +177,6 @@ class RenderPortalLayer extends RenderProxyBox {
           ),
           _ => Offset.zero,
         };
-
-    if (layer == null) {
-      layer = PortalLayer(
-        link: link,
-        showWhenUnlinked: showWhenUnlinked,
-        linkedOffset: linkedOffset,
-        unlinkedOffset: offset,
-      )..portalRenderBox = this;
-    } else {
-      layer
-        ?..link = link
-        ..showWhenUnlinked = showWhenUnlinked
-        ..linkedOffset = linkedOffset
-        ..unlinkedOffset = offset
-        ..portalRenderBox = this;
-    }
-
-    context.pushLayer(
-      layer!,
-      super.paint,
-      Offset.zero,
-      childPaintBounds: const .fromLTRB(
-        // We don't know where we'll end up, so we have no idea what our cull rect should be.
-        double.negativeInfinity,
-        double.negativeInfinity,
-        .infinity,
-        .infinity,
-      ),
-    );
-
-    assert(() {
-      layer!.debugCreator = debugCreator;
-      return true;
-    }());
-  }
-
-  @override
-  Rect get paintBounds => globalToLocal(Offset.zero) & viewSize;
-
-  @override
-  bool hitTest(BoxHitTestResult result, {required Offset position}) {
-    // Disables the hit testing if this render object is hidden.
-    if (link.childLayer == null && !showWhenUnlinked) {
-      return false;
-    }
-    // RenderPortalLayer objects don't check if they are themselves hit, because it's confusing to think about how the
-    // untransformed size and the child's transformed position interact.
-    return hitTestChildren(result, position: position);
-  }
-
-  @override
-  bool hitTestChildren(BoxHitTestResult result, {required Offset position}) => result.addWithPaintTransform(
-    transform: _currentTransform,
-    position: position,
-    hitTest: (result, position) => super.hitTestChildren(result, position: position),
-  );
-
-  @override
-  void applyPaintTransform(RenderBox child, Matrix4 transform) => transform.multiply(_currentTransform);
-
-  /// Return the transform that was used in the last composition phase, if any.
-  ///
-  /// If the [PortalLayer] has not yet been created, was never composited, or was unable to determine the transform (see
-  /// [PortalLayer.getLastTransform]), this returns the identity matrix (see [Matrix4.identity]).
-  Matrix4 get _currentTransform => layer?.getLastTransform() ?? .identity();
-
-  @override
-  void detach() {
-    notifier.removeListener(_schedule);
-    layer = null;
-    super.detach();
-  }
-
-  @override
-  bool get alwaysNeedsCompositing => true;
-
-  /// The layer we created when we were last painted.
-  @override
-  PortalLayer? get layer => super.layer as PortalLayer?;
-
-  FChangeNotifier get notifier => _notifier;
-
-  set notifier(FChangeNotifier value) {
-    if (_notifier == value) {
-      return;
-    }
-
-    _notifier.removeListener(_schedule);
-    _notifier = value;
-    _notifier.addListener(_schedule);
-  }
-
-  ChildLayerLink get link => _link;
-
-  set link(ChildLayerLink value) {
-    if (_link == value) {
-      return;
-    }
-    _link = value;
-    markNeedsPaint();
-  }
-
-  Size get viewSize => _viewSize;
-
-  set viewSize(Size value) {
-    if (_viewSize == value) {
-      return;
-    }
-
-    _viewSize = value;
-    markNeedsPaint();
-  }
-
-  /// Whether to show the render object's contents when there is no corresponding [RenderChildLayer] with the same [link].
-  ///
-  /// When the render object is linked, the child is positioned such that it has the same global position as the linked
-  /// [RenderChildLayer].
-  ///
-  /// When the render object is not linked, then: if [showWhenUnlinked] is true, the child is visible and not
-  /// repositioned; if it is false, then child is hidden, and its hit testing is also disabled.
-  bool get showWhenUnlinked => _showWhenUnlinked;
-
-  set showWhenUnlinked(bool value) {
-    if (_showWhenUnlinked == value) {
-      return;
-    }
-    _showWhenUnlinked = value;
-    markNeedsPaint();
   }
 
   /// The portal's constraints.
@@ -442,17 +264,12 @@ class RenderPortalLayer extends RenderProxyBox {
   void debugFillProperties(DiagnosticPropertiesBuilder properties) {
     super.debugFillProperties(properties);
     properties
-      ..add(DiagnosticsProperty('notifier', notifier))
-      ..add(DiagnosticsProperty('link', link))
-      ..add(DiagnosticsProperty('viewSize', viewSize))
-      ..add(DiagnosticsProperty('showWhenUnlinked', showWhenUnlinked))
       ..add(DiagnosticsProperty('portalConstraints', portalConstraints))
       ..add(DiagnosticsProperty('portalAnchor', portalAnchor))
       ..add(DiagnosticsProperty('childAnchor', childAnchor))
       ..add(DiagnosticsProperty('padding', padding))
       ..add(DiagnosticsProperty('spacing', spacing))
       ..add(ObjectFlagProperty.has('overflow', overflow))
-      ..add(DiagnosticsProperty('offset', offset))
-      ..add(TransformProperty('current transform matrix', _currentTransform));
+      ..add(DiagnosticsProperty('offset', offset));
   }
 }